Move StubStorage back into .command folder, and update JUnit tests accordingly.
